@@ -6,7 +6,7 @@
     "start": "webpack-dev-server",
     "dist": "webpack --config ./webpack-prod.config.js",
     "lint": "eslint src test",
-    "test": "mocha --compilers js:babel-core/register --require ./test/setup.js --recursive ./test",
+    "test": "mocha --require babel-core/register --recursive ./test",
     "test:watch": "npm run test -- --watch"
   },
   "author": "",
@@ -14,49 +14,43 @@
   "license": "UNLICENSED",
   "devDependencies": {
     "babel-core": "^6.18.0",
-    "babel-eslint": "^6.1.2",
+    "babel-eslint": "^10.0.1",
     "babel-loader": "^6.2.5",
     "babel-preset-airbnb": "^2.0.0",
     "babel-preset-es2015": "^6.14.0",
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-1": "^6.13.0",
     "css-loader": "^0.24.0",
-    "enzyme": "^2.4.1",
-    "eslint": "^3.4.0",
-    "eslint-config-airbnb": "^10.0.1",
-    "eslint-import-resolver-webpack": "^0.5.1",
-    "eslint-loader": "^1.5.0",
-    "eslint-plugin-import": "^1.14.0",
-    "eslint-plugin-jsx-a11y": "^2.2.0",
-    "eslint-plugin-react": "^6.2.0",
+    "eslint-plugin-prettier": "^3.0.0",
     "expect": "^1.20.2",
     "fetch-mock": "^5.1.1",
-    "jsdom": "^9.4.2",
-    "less": "^2.7.1",
+    "less": "^3.8.1",
     "less-loader": "^2.2.3",
-    "mocha": "^3.0.2",
-    "react-addons-perf": "^15.4.0",
+    "mocha": "^5.2.0",
+    "prettier": "^1.14.3",
     "react-addons-test-utils": "^15.4.0",
     "react-hot-loader": "^3.0.0-beta.6",
     "redux-devtools": "^3.3.1",
     "redux-devtools-dock-monitor": "^1.1.1",
     "redux-devtools-log-monitor": "^1.0.11",
     "redux-mock-store": "^1.1.4",
-    "style-loader": "^0.13.1",
+    "style-loader": "^0.13.2",
     "webpack": "^1.13.3",
     "webpack-dev-server": "^1.16.2"
   },
   "dependencies": {
+    "@material-ui/core": "^3.2.2",
+    "eslint": "^5.7.0",
     "immutable": "^3.8.1",
-    "material-ui": "^0.16.4",
     "moment": "^2.14.1",
-    "react": "^15.4.0",
+    "prop-types": "^15.6.2",
+    "react": "^16.5.2",
     "react-addons-pure-render-mixin": "^15.4.0",
-    "react-dom": "^15.4.0",
+    "react-dom": "^16.5.2",
     "react-redux": "^4.4.5",
-    "react-router": "^2.7.0",
+    "react-router": "^3.2.0",
     "react-router-redux": "^4.0.5",
-    "react-tap-event-plugin": "^2.0.1",
+    "react-table": "^6.8.6",
     "redux": "^3.5.2",
     "reselect": "^2.5.3",
     "whatwg-fetch": "^1.0.0"
