@@ -1,17 +1,19 @@
 // @flow
 
-import React, { Component } from 'react';
+import PropTypes from "prop-types";
+
+import React, { Component } from "react";
 
 const isValid = input => input.match(/^((\d|[1-9]\d+)(\.\d{1,2})?|\.\d{1,2})$/);
 
 class Expense extends Component {
-
   onChange = e => {
-    const newValue = e.target.value.trim() === '' ? '0' : e.target.value;
+    const newValue = e.target.value.trim() === "" ? "0" : e.target.value;
     this.props.onInputChange(
       this.props.project,
-      this.props.type === 'other' ? 'expense' : this.props.type,
-      newValue);
+      this.props.type === "other" ? "expense" : this.props.type,
+      newValue
+    );
 
     if (newValue === this.props.value.toString()) return;
     if (!isValid(newValue)) return;
@@ -19,24 +21,26 @@ class Expense extends Component {
   };
 
   render() {
-    return (<td>
-      <input
-        type='text'
-        value={this.props.input === '0' ? '' : this.props.input}
-        className={isValid(this.props.input) ? '' : 'field-error'}
-        onChange={this.onChange}
-      />
-    </td>);
+    return (
+      <td>
+        <input
+          type="text"
+          value={this.props.input === "0" ? "" : this.props.input}
+          className={isValid(this.props.input) ? "" : "field-error"}
+          onChange={this.onChange}
+        />
+      </td>
+    );
   }
 }
 
 Expense.propTypes = {
-  value: React.PropTypes.number.isRequired,
-  input: React.PropTypes.string.isRequired,
-  project: React.PropTypes.string.isRequired,
-  type: React.PropTypes.string.isRequired,
-  onValueChange: React.PropTypes.func.isRequired,
-  onInputChange: React.PropTypes.func.isRequired,
+  value: PropTypes.number.isRequired,
+  input: PropTypes.string.isRequired,
+  project: PropTypes.string.isRequired,
+  type: PropTypes.string.isRequired,
+  onValueChange: PropTypes.func.isRequired,
+  onInputChange: PropTypes.func.isRequired
 };
 
 export default Expense;
