@@ -1,9 +1,10 @@
-import React from 'react';
-import Row from './row';
+import PropTypes from "prop-types";
+import React from "react";
+import Row from "./row";
 
-const BalanceViewBody = (props) => (
+const BalanceViewBody = props => (
   <tbody>
-    {props.data.list.map(row =>
+    {props.data.list.map(row => (
       <Row
         data={row}
         onWriteOffChange={props.data.onWriteOffChange}
@@ -12,12 +13,13 @@ const BalanceViewBody = (props) => (
         onStatusChange={props.data.onStatusChange}
         onInputChange={props.data.onInputChange}
         key={row.projectId}
-      />)}
+      />
+    ))}
   </tbody>
 );
 
 BalanceViewBody.propTypes = {
-  data: React.PropTypes.object.isRequired
+  data: PropTypes.object.isRequired
 };
 
 export default BalanceViewBody;
