@@ -1,11 +1,12 @@
-import React from 'react';
-import Header from './header';
-import Body from './body';
-import Footer from './footer';
+import PropTypes from "prop-types";
+import React from "react";
+import Header from "./header";
+import Body from "./body";
+import Footer from "./footer";
 
-const BalanceViewTable = (props) => (
+const BalanceViewTable = props => (
   <div>
-    <table className='balance-table'>
+    <table className="balance-table">
       <Header />
       <Body data={props.tableData.body} />
       <Footer data={props.tableData.footer} />
@@ -14,7 +15,7 @@ const BalanceViewTable = (props) => (
 );
 
 BalanceViewTable.propTypes = {
-  tableData: React.PropTypes.object.isRequired
+  tableData: PropTypes.object.isRequired
 };
 
 export default BalanceViewTable;
