@@ -1,11 +1,21 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { getProjects, getHoursPerProject, upsertInvoiceBalance, upsertWriteOff,
-  upsertExpense, upsertStatus, changeInput } from '../actions/index';
-import titleSelector from '../selectors/titleSelector';
-import tableBodySelector from '../selectors/tableBodySelector';
-import footerSelector from '../selectors/footerSelector';
-import IndexComponent from '../components/index';
+import PropTypes from "prop-types";
+import React, { Component } from "react";
+import { connect } from "react-redux";
+import {
+  getProjects,
+  getHoursPerProject,
+  upsertInvoiceBalance,
+  upsertWriteOff,
+  upsertExpense,
+  upsertStatus,
+  changeInput
+} from "../actions/index";
+import { initateTimeTrackingReportDownload } from "../apiclient";
+import titleSelector from "../selectors/titleSelector";
+import tableBodySelector from "../selectors/tableBodySelector";
+import footerSelector from "../selectors/footerSelector";
+import filterFieldsSelector from "../selectors/filterFieldsSelector";
+import IndexComponent from "../components/index";
 
 class App extends Component {
   constructor(props) {
@@ -15,69 +25,102 @@ class App extends Component {
     props.getHoursPerProject(props.title.startDate, props.title.endDate);
   }
 
+  componentDidMount() {
+    window.addEventListener("resize", () => this.forceUpdate());
+  }
+
   componentWillReceiveProps(nextProps) {
-    if (nextProps.title.startDate !== this.props.title.startDate ||
-        nextProps.title.endDate !== this.props.title.endDate) {
+    if (
+      nextProps.title.startDate !== this.props.title.startDate ||
+      nextProps.title.endDate !== this.props.title.endDate
+    ) {
       nextProps.getHoursPerProject(nextProps.title.startDate, nextProps.title.endDate);
     }
   }
 
-  onWriteOffChange = (project, minutes) =>
-     this.props.upsertWriteOff(project, this.props.title.endDate, minutes);
+  onWriteOffChange = (project, hours) =>
+    this.props.upsertWriteOff(project, this.props.title.endDate, hours * 60);
 
-  onExpenseChange = (project, type, minutes) =>
-     this.props.upsertExpense(project, this.props.title.endDate, type, minutes);
+  onOtherExpenseChange = (project, money) =>
+    this.props.upsertExpense(project, this.props.title.endDate, "other", money);
 
-  onInvoiceBalanceChange = (project, minutes, money) =>
+  onSubcontractorExpenseChange = (project, money) =>
+    this.props.upsertExpense(project, this.props.title.endDate, "subcontractor", money);
+
+  onInvoiceBalanceMinutesChange = (project, hours) => {
+    const money = this.props.tableBody.data.find(row => row.projectId === project)["invoice_money"];
+    this.props.upsertInvoiceBalance(project, this.props.title.endDate, hours * 60, money);
+  };
+
+  onInvoiceBalanceMoneyChange = (project, money) => {
+    const minutes = this.props.tableBody.data.find(row => row.projectId === project)[
+      "invoice_minutes"
+    ];
     this.props.upsertInvoiceBalance(project, this.props.title.endDate, minutes, money);
+  };
 
   onStatusChange = (project, status) =>
     this.props.upsertStatus(project, this.props.title.endDate, status);
 
-  onInputChange = (project, key, value) =>
-    this.props.changeInput(project, key, value);
+  onInputChange = (project, key, value) => this.props.changeInput(project, key, value);
+
+  getMonthlyTimeTrackingReport = projectId => {
+    initateTimeTrackingReportDownload(
+      projectId,
+      this.props.title.startDate,
+      this.props.title.endDate
+    );
+  };
 
   render() {
+    // this.props.tableBody.data.map(x => console.log(x));
     if (this.props.tableBody.loading) {
       return null;
     }
-    return (<IndexComponent
-      tableData={{
-        body: {
-          list: this.props.tableBody.data,
-          onWriteOffChange: this.onWriteOffChange,
-          onExpenseChange: this.onExpenseChange,
-          onInvoiceBalanceChange: this.onInvoiceBalanceChange,
-          onStatusChange: this.onStatusChange,
-          onInputChange: this.onInputChange,
-        },
-        footer: this.props.footer.data,
-      }}
-      title={this.props.title}
-    />);
+    return (
+      <IndexComponent
+        tableData={{
+          body: {
+            list: this.props.tableBody.data,
+            filterFieldValues: this.props.filterFieldValues,
+            onWriteOffChange: this.onWriteOffChange,
+            onOtherExpenseChange: this.onOtherExpenseChange,
+            onSubcontractorExpenseChange: this.onSubcontractorExpenseChange,
+            onInvoiceBalanceMinutesChange: this.onInvoiceBalanceMinutesChange,
+            onInvoiceBalanceMoneyChange: this.onInvoiceBalanceMoneyChange,
+            onStatusChange: this.onStatusChange,
+            onInputChange: this.onInputChange,
+            getMonthlyTimeTrackingReport: this.getMonthlyTimeTrackingReport
+          },
+          footer: this.props.footer.data
+        }}
+        title={this.props.title}
+      />
+    );
   }
 }
 
 App.propTypes = {
   // mapStateToProps
-  tableBody: React.PropTypes.object.isRequired,
-  title: React.PropTypes.object.isRequired,
-  footer: React.PropTypes.object.isRequired,
+  tableBody: PropTypes.object.isRequired,
+  title: PropTypes.object.isRequired,
+  footer: PropTypes.object.isRequired,
 
   // mapDispatchToProps
-  getProjects: React.PropTypes.func.isRequired,
-  getHoursPerProject: React.PropTypes.func.isRequired,
-  upsertInvoiceBalance: React.PropTypes.func.isRequired,
-  upsertWriteOff: React.PropTypes.func.isRequired,
-  upsertExpense: React.PropTypes.func.isRequired,
-  upsertStatus: React.PropTypes.func.isRequired,
-  changeInput: React.PropTypes.func.isRequired,
+  getProjects: PropTypes.func.isRequired,
+  getHoursPerProject: PropTypes.func.isRequired,
+  upsertInvoiceBalance: PropTypes.func.isRequired,
+  upsertWriteOff: PropTypes.func.isRequired,
+  upsertExpense: PropTypes.func.isRequired,
+  upsertStatus: PropTypes.func.isRequired,
+  changeInput: PropTypes.func.isRequired
 };
 
 const mapStateToProps = (state, ownProps) => ({
   tableBody: tableBodySelector(state, ownProps),
   title: titleSelector(state, ownProps),
   footer: footerSelector(state, ownProps),
+  filterFieldValues: filterFieldsSelector(state, ownProps)
 });
 
 const mapDispatchToProps = {
@@ -87,7 +130,10 @@ const mapDispatchToProps = {
   upsertWriteOff,
   upsertExpense,
   upsertStatus,
-  changeInput,
+  changeInput
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(App);
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(App);
