@@ -1,9 +1,10 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import Dialog from 'material-ui/Dialog';
-import FlatButton from 'material-ui/FlatButton';
+import PropTypes from "prop-types";
+import React, { Component } from "react";
+import { connect } from "react-redux";
+import Dialog from "@material-ui/core/Dialog";
+import FlatButton from "@material-ui/core/FlatButton";
 
-import { clearApiError } from '../actions';
+import { clearApiError } from "../actions";
 
 class ErrorDialog extends Component {
   handleClose = () => {
@@ -12,18 +13,13 @@ class ErrorDialog extends Component {
 
   render() {
     const actions = [
-      <FlatButton
-        label='Got it'
-        primary
-        keyboardFocused
-        onTouchTap={this.handleClose}
-      />,
+      <FlatButton label="Got it" primary keyboardFocused onTouchTap={this.handleClose} />
     ];
 
     return (
       <div>
         <Dialog
-          title='Error'
+          title="Error"
           actions={actions}
           modal
           open={this.props.error !== null}
@@ -37,12 +33,15 @@ class ErrorDialog extends Component {
 }
 
 ErrorDialog.propTypes = {
-  error: React.PropTypes.string,
-  clearApiError: React.PropTypes.func
+  error: PropTypes.string,
+  clearApiError: PropTypes.func
 };
 
-const mapStateToProps = (state) => ({
+const mapStateToProps = state => ({
   error: state.error
 });
 
-export default connect(mapStateToProps, { clearApiError })(ErrorDialog);
+export default connect(
+  mapStateToProps,
+  { clearApiError }
+)(ErrorDialog);
