@@ -1,5 +1,5 @@
-import { createSelector } from 'reselect';
-import * as Immutable from 'immutable';
+import { createSelector } from "reselect";
+import * as Immutable from "immutable";
 
 const grossTurnoverCustomers = (projects, hoursPerProject) => {
   if (projects.loading || hoursPerProject.loading) {
@@ -7,12 +7,10 @@ const grossTurnoverCustomers = (projects, hoursPerProject) => {
   }
   return {
     loading: false,
-    data: hoursPerProject.data.reduce((result, value, key) =>
-      result.update(
-        projects.data.get(key).customer.id,
-        0,
-        x => x + value.invoice_balance_money
-      ), new Immutable.Map()
+    data: hoursPerProject.data.reduce(
+      (result, value, key) =>
+        result.update(projects.data.get(key).customer.id, 0, x => x + value.invoice_balance_money),
+      new Immutable.Map()
     )
   };
 };
@@ -20,5 +18,5 @@ const grossTurnoverCustomers = (projects, hoursPerProject) => {
 export default createSelector(
   state => state.projects,
   state => state.hours_per_project,
-  grossTurnoverCustomers,
+  grossTurnoverCustomers
 );
