@@ -1,16 +1,16 @@
-const webpack = require('webpack');
+const webpack = require("webpack");
 
 module.exports = {
-  entry: ['./src/index.js'],
+  entry: ["./src/index.js"],
   output: {
     path: `${__dirname}/dist/js`,
-    filename: 'app.bundle.js'
+    filename: "app.bundle.js"
   },
-  devtool: 'source-map',
+  devtool: "source-map",
   plugins: [
     new webpack.DefinePlugin({
-      'process.env': {
-        NODE_ENV: JSON.stringify('production')
+      "process.env": {
+        NODE_ENV: JSON.stringify("production")
       }
     }),
     new webpack.optimize.UglifyJsPlugin({
@@ -22,8 +22,9 @@ module.exports = {
   ],
   module: {
     loaders: [
-        { test: /\.less$/, loader: 'style!css!less' },
-        { test: /\.js$/, loaders: ['babel'], exclude: /node_modules/, include: __dirname }
+      { test: /\.less$/, loader: "style!css!less" },
+      { test: /\.js$/, loaders: ["babel"], exclude: /node_modules/, include: __dirname },
+      { test: /\.css$/, loader: "style!css" }
     ]
   }
 };
